+++ Zolian.Server.Base/GameScripts/Skills/ClericSkillTree.cs
@@ -229,10 +229,10 @@ public class Smite : SkillScript
         }
     }
 
-    private int DamageCalc(Sprite sprite)
+    private long DamageCalc(Sprite sprite)
     {
         _crit = false;
-        int dmg;
+        long dmg;
         if (sprite is Aisling damageDealingAisling)
         {
             var client = damageDealingAisling.Client;
